Upload dataset
I have download the database of our interest that has been approved by our assigned TA.
While I was on kaggle, I also discovered the extension of the database with 2019 datas in it. I have attached it that we might be able to make use of it.
Link is here for reference: https://www.kaggle.com/ashaheedq/video-games-sales-2019